LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-items-percentage-margins-002-expected.txt


PASS .grid 1 
PASS .grid 2 
FAIL .grid 3 assert_equals: 
<div class="grid directionRTL">
  <div class="marginRight50Percent" data-expected-margin-right="50" data-offset-x="440" data-expected-width="10" data-expected-height="10">X</div>
  <div data-offset-x="400" data-offset-y="10" data-expected-width="100" data-expected-height="10"></div>
</div>
offsetLeft expected 440 but got 365
FAIL .grid 4 assert_equals: 
<div class="grid directionRTL">
  <div class="marginLeft50Percent" data-expected-margin-left="50" data-offset-x="490" data-expected-width="10" data-expected-height="10">X</div>
  <div data-offset-x="400" data-offset-y="10" data-expected-width="100" data-expected-height="10"></div>
</div>
margin-left expected "50" but got "125"
PASS .grid 5 
PASS .grid 6 
FAIL .grid 7 assert_equals: 
<div class="grid">
  <div class="marginRight50Percent" data-expected-margin-right="50" data-offset-x="0" data-expected-width="10" data-expected-height="10">X</div>
  <div data-offset-x="0" data-offset-y="10" data-expected-width="100" data-expected-height="10"></div>
</div>
margin-right expected "50" but got "125"
FAIL .grid 8 assert_equals: 
<div class="grid">
  <div class="marginLeft50Percent" data-expected-margin-left="50" data-offset-x="50" data-expected-width="10" data-expected-height="10">X</div>
  <div data-offset-x="0" data-offset-y="10" data-expected-width="100" data-expected-height="10"></div>
</div>
offsetLeft expected 50 but got 125
Direction LTR

Item margin-left: 50%;
X
Item margin-right: 50%;
X
Item margin-top: 50%;
X
Item margin-bottom: 50%;
X
Direction RTL

Item margin-left: 50%;
X
Item margin-right: 50%;
X
Item margin-top: 50%;
X
Item margin-bottom: 50%;
X
